Add files via upload
Fundamental matrix serves as a basic unit for computer vision and stereo geometry. Fundamental matrix is a 3 x 3 matrix of rank 2. Its described by:
X . [F]. X’ = 0
We can infer epipolar lines, epipolar plane, epipoles of the image taken by two cameras from different points through fundamental matrix. We first compute SIFT then matches, follow the same procedure we followed for computing homography matrix and then we can compute fundamental matrix with opencv.

Epilines: Epilines are the group of points that are projected .The line OL–X is seen by the left camera as a point because it is directly in line with that camera's lens optical center. However, the right camera sees this line as a line in its image plane. That line (eR–xR) in the right camera is called an epipolar line. Ref: wiki Keypoint matching and fundamental matrix is necessary for computing epilines.

 Disparity:
Computed disparity of the images with built in function of opencv by StreaoBM(), for supporting cv2, we are using StereoBM_create